fold into existing restore/save_build_cache
